requirements.txt
amqp
asgiref
Babel
backcall
billiard
celery
certifi
cffi
chardet
click
click-didyoumean
click-plugins
click-repl
coreapi
coreschema
cryptography
decorator
defusedxml
Django
django-cors-headers
django-filter
djangorestframework
drf-yasg
flower
future
humanize
idna
importlib-metadata
inflection
ipython
ipython-genutils
itypes
jedi
Jinja2
kombu
MarkupSafe
oauthlib
openapi-codec
packaging
parso
pexpect
pickleshare
Pillow
prometheus-client
prompt-toolkit
psycopg2-binary
ptyprocess
pycparser
Pygments
PyJWT
pyparsing
python3-openid
pytz
requests
requests-oauthlib
ruamel.yaml
ruamel.yaml.clib
simplejson
six
sqlparse
tornado
traitlets
uritemplate
urllib3
uWSGI
vine
wcwidth
zipp
